separated modals to modal component
